requirements.txt
python-dotenv~=1.0.0
uvicorn~=0.15.0
PyGithub~=1.58.1
schedule~=1.2.0
pydantic~=1.10.7
fastapi~=0.70.1